Fałszywy alarm wszystko OK

+++ zajecia/Miziol/projekt/odwrocona_notacja_polska.cpp
@@ -15,7 +15,7 @@ struct stos
 		{
 		return p==0;
 		}
-	void insert(int r)
+	void insert(float r)
 		{
 		t[p]=r;
 		p++;
